output/prototypes/ai_video_pipeline/requirements.txt
groq>=0.9.0
moviepy>=1.0.3
edge-tts>=6.1.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
